% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\alias{merge_anno}
\alias{merge_context}
\title{Merge tsv outputs from \code{\link{summarise}} functions into one file.}
\usage{
merge_anno(anno_path, force = FALSE)

merge_context(context_path, force = FALSE)
}
\arguments{
\item{anno_path}{path to the directory containing the summarised tsv files for each cell for a given annotation}

\item{force}{Logical, whether to overwrite even if the output file exists}

\item{context_path}{path to the directory containing the summarised tsv files for each annotation for a given context}
}
\value{
Writes merged tsv files to the same directory as the annotation/context files and returns TRUE
}
\description{
These functions merge the tsv files in a given path into a single tsv file.
}
