% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reducedDim.R
\name{plot_reducedDim}
\alias{plot_reducedDim}
\title{Plot reduced dimensions for SCE object}
\usage{
plot_reducedDim(
  sce,
  reducedDim = "PCA",
  comps = c(1, 2),
  colBy = "Batch",
  facet_by = NULL,
  cols = NULL,
  axis.title = "PC",
  ...
)
}
\arguments{
\item{sce}{A SCE object}

\item{reducedDim}{Name of the reduced dimension from \code{names(sce@int_colData$reducedDims)}}

\item{comps}{Integer vector of length 2. The components to use.}

\item{colBy}{Character or integer. colData to use to colour the cells. From \code{names(colData(sce))}}

\item{facet_by}{Character or integer. colData to use to facet the plots. From \code{names(colData(sce))}}

\item{cols}{Named character of colours to use. If NULL default values are generated.}

\item{axis.title}{Title prefix of the 2 axes, e.g. 'PC' or 'UMAP' etc.}

\item{...}{Additional parameters passed to \code{\link[ggplot2]{geom_point}}}
}
\value{
A ggplot2 object
}
\description{
Plot reduced dimensions for SCE object
}
