% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\alias{summarise_sample}
\alias{summarise_anno}
\alias{summarise_context}
\title{Quantify DNA methylation and chromatin accessibility over genomic features}
\usage{
summarise_sample(
  filepath,
  annotation_file,
  anno_name,
  context = c("CG", "GC"),
  valid_chromosomes = as.character(c(1:22, "MT", "X", "Y")),
  a = 0,
  b = 0,
  out.dir,
  force = FALSE
)

summarise_anno(
  samplefiles.path,
  annotation_file,
  anno_name,
  context = c("CG", "GC"),
  valid_chromosomes = as.character(c(1:22, "MT", "X", "Y")),
  a = 0,
  b = 0,
  out.dir,
  cores = parallel::detectCores(),
  force = FALSE
)

summarise_context(
  samplefiles.path,
  annotation_files,
  context = c("CG", "GC"),
  valid_chromosomes = as.character(c(1:22, "MT", "X", "Y")),
  a = 0,
  b = 0,
  out.dir,
  cores = parallel::detectCores(),
  force = FALSE
)

summarise(
  datapath = "data",
  annotation_files,
  valid_chromosomes = as.character(c(1:22, "MT", "X", "Y")),
  a = 0,
  b = 0,
  force,
  cores = parallel::detectCores(),
  out.dir = "output/scnomeseq"
)
}
\arguments{
\item{filepath}{path to cov.gz file}

\item{annotation_file}{path to the annotation file that has \code{c("chr","start","end","strand","id")} columns}

\item{anno_name}{Character, annotation name to use and add to the summarised table, e.g. 'Enhancer'}

\item{context}{One of c('CG', 'GC')}

\item{valid_chromosomes}{Character vector of chromosomes to keep. Must match the chromosome notation in the data}

\item{a, }{b Beta prior parameters. See SN 1 \link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4117646/}}

\item{out.dir}{Output directory to save the tsv file. The output file will be of form: outdir/CpG_or_GpC/anno_name/cov_file_name.tsv}

\item{force}{Logical, whether to overwrite even if the output file exists}

\item{samplefiles.path}{path to all cov.gz files for a given context}

\item{cores}{Integer, number of CPUS to use}

\item{annotation_files}{Named character vector (name = 'path') where names will be used as annotation names}

\item{datapath}{Path to the data. It includes datapath/CpG and datapath/GpC subdirectories which contain the cov files for all cells}
}
\value{
Writes file to out.dir and returns NULL
}
\description{
These functions overlap the individual CpG sites with genomic features such as
promoters, gene bodies, etc.
}
\details{
\itemize{
\item{\code{summarise_sample} summarises the calls for a given cov.gz file and annotation file}
\item{\code{summarise_anno} summarises the calls for a given annotation file for all cells}
\item{\code{summarise_context} summarises the calls for a all annotation file for all cells for a given context (CG or GC)}
\item{\code{summarise} summarises the calls for a all annotation file for all cells for both contexts (CG and GC)}
}
}
