% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{cat2}
\alias{cat3}
\alias{cat4}
\alias{rm_trailing}
\alias{file_exists}
\alias{check_file_exists}
\title{Utilities}
\usage{
cat2(fmt, ...)

cat3(fmt, ...)

cat4(fmt, ...)

rm_trailing(path)

file_exists(filepath)

check_file_exists(filepath)
}
\arguments{
\item{fmt, ...}{passed to \code{\link[base]{sprintf}}}

\item{filepath, path}{path to file}
}
\value{
None
}
\description{
Utility functions
}
\examples{
cat2("this is a '\%s' text", 'yellow')
cat3("this is a '\%s' text", 'purple')
cat4("this is a '\%s' text", 'blue')
rm_trailing("foo/bar/")
# "foo/bar"
}
